+++ example/lib/main.dart
@@ -17,7 +17,7 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   final _flutterPinelabsPlugin = const FlutterPinelabs(
     header: HeaderModel(
-      applicationId: 'your application id.',
+      applicationId: '43640818b0cd432498aa166d5943f591',
       methodId: '1001',
       versionNo: '1.0',
     ),
@@ -50,6 +50,145 @@ class _MyAppState extends State<MyApp> {
     );
 
     /// provides ResponseModel in return which contains the response from the pinelabs device.
+    setState(() {
+      _responseMessage = response?.responseMsg ?? '';
+    });
+  }
+
+  String _organizeName(String name) {
+    final words = name.split(" ");
+    StringBuffer line = StringBuffer();
+    final result = StringBuffer();
+
+    for (var i = 0; i < words.length; i++) {
+      final word = words[i];
+      if (line.length + word.length > 16) {
+        //add line to result if it full
+        result
+          ..write(line)
+          ..write('\n');
+        line = StringBuffer(); //reset line is empty
+      }
+      line.write(word);
+      if (i == (words.length - 1)) {
+        line.write(" ".padRight(line.length >= 16 ? 0 : (16 - words.length)));
+      } else {
+        line.write(" ");
+      }
+    }
+    result.write(line);
+
+    return result.toString();
+  }
+
+  Future<void> _printData() async {
+    final rowData = [
+      {'name': 'test product', 'qty': '2', 'price': '\$200'},
+      {'name': 'dosa', 'qty': '15', 'price': '\$10'},
+      {
+        'name': 'SOME BRAND awesome product that keeps your mind cool',
+        'qty': '2',
+        'price': '\$200'
+      },
+      {'name': 'test product', 'qty': '100', 'price': '\$10000'},
+    ];
+    final List<PrintRowModel> printRows = rowData.map((e) {
+      final name = e['name']!;
+      final qty = e['qty']!;
+      final price = e['price']!;
+      e['name'] = name.length > 12 ? _organizeName(name) : name.padRight(16);
+      e['qty'] = qty.padRight(7);
+      e['price'] = price.padRight(9);
+      return PrintRowModel(
+        printDataType: '0',
+        printerWidth: 32,
+        dataToPrint: '${e['name']}${e['qty']}${e['price']}',
+        imageData: '0',
+        imagePath: '0',
+        isCenterAligned: false,
+      );
+    }).toList();
+
+    final request = PrintModel(
+      printRefNo: '1234',
+      savePrintData: true,
+      data: [
+        const PrintRowModel(
+          printDataType: '0',
+          printerWidth: 24,
+          isCenterAligned: true,
+          dataToPrint: 'Test Bill',
+          imageData: '0',
+          imagePath: '0',
+        ),
+        const PrintRowModel(
+          printDataType: '0',
+          printerWidth: 48,
+          dataToPrint: '------------------------------------------------',
+          imageData: '0',
+          imagePath: '0',
+        ),
+        const PrintRowModel(
+          printDataType: '0',
+          printerWidth: 24,
+          dataToPrint: 'Items     Qty     Cost',
+          imageData: '0',
+          imagePath: '0',
+        ),
+        const PrintRowModel(
+          printDataType: '0',
+          printerWidth: 48,
+          dataToPrint: '------------------------------------------------',
+          imageData: '0',
+          imagePath: '0',
+        ),
+        ...printRows,
+        const PrintRowModel(
+          printDataType: '0',
+          printerWidth: 48,
+          dataToPrint: '------------------------------------------------',
+          imageData: '0',
+          imagePath: '0',
+        ),
+        const PrintRowModel(
+          printDataType: '0',
+          printerWidth: 24,
+          dataToPrint: 'Total:     \$5000',
+          imageData: '',
+          imagePath: '0',
+        ),
+        const PrintRowModel(
+          printDataType: '0',
+          printerWidth: 48,
+          dataToPrint: '------------------------------------------------',
+          imageData: '0',
+          imagePath: '0',
+        ),
+        const PrintRowModel(
+          printDataType: '0',
+          printerWidth: 32,
+          dataToPrint: 'Thank You For Shopping With Us',
+          imageData: '',
+          imagePath: '0',
+        ),
+        const PrintRowModel(
+          printDataType: '0',
+          printerWidth: 24,
+          dataToPrint: '\n',
+          imageData: '',
+          imagePath: '0',
+        ),
+      ],
+    );
+    final response = await _flutterPinelabsPlugin.printData(
+      overrideHeader: const HeaderModel(
+        applicationId: '3b3dabce3af34173a4c6d95e4f696e15',
+        methodId: '1002',
+        versionNo: '1.0',
+      ),
+      printRequest: request,
+    );
+
     setState(() {
       _responseMessage = response?.responseMsg ?? '';
     });
@@ -89,10 +228,17 @@ class _MyAppState extends State<MyApp> {
             ),
             const SizedBox(height: 20),
             ElevatedButton(
-              child: const Text('Send request'),
+              child: const Text('Do transaction'),
               onPressed: () async {
                 double amount = double.tryParse(_controller.text) ?? 0;
                 _doTransaction(amount);
+              },
+            ),
+            const SizedBox(height: 20),
+            ElevatedButton(
+              child: const Text('Print Data'),
+              onPressed: () async {
+                _printData();
               },
             ),
             const SizedBox(height: 20),