+++ lib/models/models.dart
@@ -1,3 +1,5 @@
 export 'header_model.dart';
+export 'print_model.dart';
+export 'print_row_model.dart';
 export 'response_model.dart';
 export 'transaction_model.dart';