+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 * Feature implemented:  
   * doTransaction
   * sendRequest
+
+## 0.1.0+1
+
+* added print data model
+* added example to print receipt using string.